pyproject.toml
[project]
name = "student-dropout-mlops"
version = "0.1.0"
description = "Production-ready MLOps stack for predicting student dropout in African higher education."
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
authors = [{ name = "UZ MSc Computer Science" }]

dependencies = [
  # API
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "pydantic-settings>=2.4",
  "python-multipart>=0.0.9",
  "slowapi>=0.1.9",
  "PyJWT>=2.9",
  "passlib[bcrypt]>=1.7.4",

  # Logging / metrics
  "structlog>=24.1",
  "prometheus-client>=0.20",

  # Data + ML
  "pandas>=2.2",
  "numpy>=1.26,<2.1",
  "scikit-learn>=1.5",
  "xgboost>=2.1",
  "lightgbm>=4.5",
  "torch>=2.4",
  "joblib>=1.4",
  "pyarrow>=17.0",

  # Validation, tracking, drift
  "pandera[pandas]>=0.20",
  "mlflow>=2.16",
  "evidently>=0.4.33",
  "shap>=0.46",

  # Pipeline orchestration
  "dvc>=3.55",
  "pyyaml>=6.0",
  "requests>=2.32",
  "ucimlrepo>=0.0.7",
  "matplotlib>=3.9",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6",
  "mypy>=1.11",
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "pytest-asyncio>=0.24",
  "httpx>=0.27",
  "types-requests",
  "types-PyYAML",
  "pandas-stubs",
  "pre-commit>=3.8",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]

# ---------------------------------------------------------------------
# Ruff: lint + format (replaces black/isort/flake8)
# ---------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["notebooks", "frontend", "data", "models", "mlruns", "reports"]

[tool.ruff.lint]
select = [
  "E", "F", "W",   # pycodestyle / pyflakes
  "I",              # isort
  "UP",             # pyupgrade
  "B",              # bugbear
  "C4",             # comprehensions
  "SIM",            # simplify
  "RUF",            # ruff-specific
  "N",              # naming
  "PL",             # pylint subset
  "TID",            # tidy imports
]
ignore = ["PLR0913", "PLR2004"]

[tool.ruff.lint.per-file-ignores]
"backend/tests/*" = ["S101", "PLR2004"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ---------------------------------------------------------------------
# Mypy: strict-ish typing
# ---------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_decorators = false
plugins = ["pydantic.mypy"]
exclude = ["notebooks", "frontend", "build", "dist"]

[[tool.mypy.overrides]]
module = [
  "shap.*", "xgboost.*", "lightgbm.*", "evidently.*",
  "mlflow.*", "sklearn.*", "joblib.*", "pandera.*",
  "ucimlrepo.*", "slowapi.*", "matplotlib.*",
]
ignore_missing_imports = true

# ---------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------
[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
  "-ra",
  "--strict-markers",
  "--strict-config",
  "--cov=backend/app",
  "--cov-report=term-missing",
  "--cov-report=xml",
]
testpaths = ["backend/tests"]
asyncio_mode = "auto"
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
branch = true
source = ["backend/app"]
omit = [
  "backend/app/ml/models/mlp.py",  # PyTorch training is integration-tested, not unit
]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "if __name__ == .__main__.:",
  "raise NotImplementedError",
  "if TYPE_CHECKING:",
]
